+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix the server-side `Request` class implementation
 * Fix `Pledge` instances executing the callbacks synchronously
 * Fix `Cache#has(key)` return true for expired entries
+* Fix `memoize` Decorator not setting `max_age` when `ignore_callbacks` is true
 
 ## 0.6.3 (2018-12-06)